Morning of 7/10 - need to fix layout of people list on Dashboard.

+++ JobHuntingApp/JobHuntingApp/Controllers/CoverLettersController.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using JobHuntingApp.Data;
@@ -10,31 +12,48 @@ using JobHuntingApp.Models;
 
 namespace JobHuntingApp.Controllers
 {
+    [Authorize]
     public class CoverLettersController : Controller
     {
         private readonly JobHuntContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
+        private readonly SignInManager<ApplicationUser> _signInManager;
 
-        public CoverLettersController(JobHuntContext context)
+        public CoverLettersController(JobHuntContext context, UserManager<ApplicationUser> userManager,
+          SignInManager<ApplicationUser> signInManager)
         {
-            _context = context;    
+            _context = context;
+            _userManager = userManager;
+            _signInManager = signInManager;
         }
 
         // GET: CoverLetters
         public async Task<IActionResult> Index()
         {
-            return View(await _context.CoverLetters.ToListAsync());
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
+            return View(await _context.CoverLetters.Where(x => x.UserID == user.Id).ToListAsync());
         }
 
         // GET: CoverLetters/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var coverLetter = await _context.CoverLetters
-                .SingleOrDefaultAsync(m => m.CoverLetterID == id);
+            var coverLetter = await _context.CoverLetters.SingleOrDefaultAsync(m => m.CoverLetterID == id && m.UserID==user.Id);
             if (coverLetter == null)
             {
                 return NotFound();
@@ -56,8 +75,15 @@ namespace JobHuntingApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("CoverLetterID,JobID,CoverLetterText")] CoverLetter coverLetter)
         {
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
             if (ModelState.IsValid)
             {
+                coverLetter.UserID = user.Id;
                 _context.Add(coverLetter);
                 await _context.SaveChangesAsync();
                 return RedirectToAction("Index");
@@ -68,12 +94,18 @@ namespace JobHuntingApp.Controllers
         // GET: CoverLetters/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var coverLetter = await _context.CoverLetters.SingleOrDefaultAsync(m => m.CoverLetterID == id);
+            var coverLetter = await _context.CoverLetters.SingleOrDefaultAsync(m => m.CoverLetterID == id && m.UserID==user.Id);
             if (coverLetter == null)
             {
                 return NotFound();
@@ -88,6 +120,12 @@ namespace JobHuntingApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("CoverLetterID,JobID,CoverLetterText")] CoverLetter coverLetter)
         {
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
             if (id != coverLetter.CoverLetterID)
             {
                 return NotFound();
@@ -97,6 +135,7 @@ namespace JobHuntingApp.Controllers
             {
                 try
                 {
+                    coverLetter.UserID = user.Id;
                     _context.Update(coverLetter);
                     await _context.SaveChangesAsync();
                 }
@@ -119,13 +158,19 @@ namespace JobHuntingApp.Controllers
         // GET: CoverLetters/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
             var coverLetter = await _context.CoverLetters
-                .SingleOrDefaultAsync(m => m.CoverLetterID == id);
+                .SingleOrDefaultAsync(m => m.CoverLetterID == id && m.UserID==user.Id);
             if (coverLetter == null)
             {
                 return NotFound();
@@ -139,7 +184,13 @@ namespace JobHuntingApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var coverLetter = await _context.CoverLetters.SingleOrDefaultAsync(m => m.CoverLetterID == id);
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
+            var coverLetter = await _context.CoverLetters.SingleOrDefaultAsync(m => m.CoverLetterID == id && m.UserID==user.Id);
             _context.CoverLetters.Remove(coverLetter);
             await _context.SaveChangesAsync();
             return RedirectToAction("Index");
@@ -148,6 +199,11 @@ namespace JobHuntingApp.Controllers
         private bool CoverLetterExists(int id)
         {
             return _context.CoverLetters.Any(e => e.CoverLetterID == id);
+        }
+
+        private Task<ApplicationUser> GetCurrentUserAsync()
+        {
+            return _userManager.GetUserAsync(HttpContext.User);
         }
     }
 }
